+++ CHANGELOG.md
@@ -1,12 +1,13 @@
 ## 2.0.0 (June 13, 2019)
 
 - BREAKING: Changed default class for results list from `autocomplete-results` to `autocomplete-result-list`
+- BREAKING: Changed default ID for results list from `autocomplete-results-{id}` to `autocomplete-result-list-{id}`
 - BREAKING: Call `search` function on focus
 
 ### autocomplete-vue
 
-- BREAKING: Renamed old default slot to "results"
-- BREAKING: Added new default slot to control rendering of entire component
+- BREAKING: Updated default slot to provide template for entire component
+- BREAKING: Added new slot called "result" to provide template for a single result item
 - BREAKING: Removed `onSubmit` prop to use custom `submit` event instead
 
 ## 1.0.2 (May 1, 2019)